chore(release): bump version and deploy packages (#199)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.

# Releases
## @mheob/tsconfig@2.1.1

### Patch Changes

-   [#198](https://github.com/mheob/config/pull/198) ([@mheob](https://github.com/mheob)): Fix the path to the `tsconfig.json` files.